Rename composables file to components
