+++ app/src/main/java/com/dubproductions/bracket/ui/main/hosting/EditTournament.kt
@@ -19,7 +19,6 @@ import androidx.compose.material3.Text
 import androidx.compose.material3.TextButton
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.stringResource
@@ -29,7 +28,7 @@ import androidx.compose.ui.unit.sp
 import com.dubproductions.bracket.R
 import com.dubproductions.bracket.data.Tournament
 import com.dubproductions.bracket.data.TournamentStatus
-import com.dubproductions.bracket.ui.ReusableDialog
+import com.dubproductions.bracket.ui.components.ReusableDialog
 import com.dubproductions.bracket.ui_state.EditTournamentUIState
 import java.text.SimpleDateFormat
 import java.util.Locale
