+++ app/src/main/java/com/dubproductions/bracket/ui/main/hosting/Hosting.kt
@@ -11,14 +11,13 @@ import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.FloatingActionButton
 import androidx.compose.material3.Icon
 import androidx.compose.material3.Scaffold
-import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import com.dubproductions.bracket.R
 import com.dubproductions.bracket.data.Tournament
-import com.dubproductions.bracket.ui.TournamentSummaryCard
+import com.dubproductions.bracket.ui.components.TournamentSummaryCard
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
