+++ app/src/main/java/com/dubproductions/bracket/ui/main/hosting/TournamentCreation.kt
@@ -22,15 +22,13 @@ import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
 import com.dubproductions.bracket.R
 import com.dubproductions.bracket.Type
 import com.dubproductions.bracket.Validation
-import com.dubproductions.bracket.ui.ReusableDialog
+import com.dubproductions.bracket.ui.components.ReusableDialog
 import com.dubproductions.bracket.ui_state.TournamentCreationUIState
 
 @Composable
